requirements.txt
flask
flask-cors
opencv-python-headless
numpy
Pillow
